+++ run-linux.sh
@@ -22,7 +22,7 @@ Electric Book options
 1. Create a print PDF
 2. Create a screen PDF
 3. Run as a website
-4. Create EPUB-ready files
+4. Create an epub
 5. Export to Word
 6. Install or update dependencies
 7. Exit
@@ -95,7 +95,7 @@ If not, just hit return."
 			fi
 			# Run prince, showing progress (-v), printing the docs in file-list
 			# and saving the resulting PDF to the _output folder
-			prince -v -l file-list -o ../../../_output/$printpdffilename.pdf
+			prince -v -l file-list -o ../../../_output/$printpdffilename.pdf --javascript
 			# Navigate back to where we began.
 			cd "$location"
 			# Tell the user we're done
@@ -180,7 +180,7 @@ If not, just hit return."
 			fi
 			# Run prince, showing progress (-v), printing the docs in file-list
 			# and saving the resulting PDF to the _output folder
-			prince -v -l file-list -o ../../../_output/$screenpdffilename.pdf
+			prince -v -l file-list -o ../../../_output/$screenpdffilename.pdf --javascript
 			# Navigate back to where we began.
 			cd "$location"
 			# Tell the user we're done
@@ -266,13 +266,9 @@ You may need to reload the web page once this server is running."
 		# Ask if we're outputting the files from a subdirectory (e.g. a translation)
 		echo "If you're outputting files in a subdirectory (e.g. a translation), type its name. Otherwise, hit enter. "
 		read epubsubdirectory
-		echo -n "What is the first file in your book? Usually the cover.
-(Don't include the file extension. Hit enter for the default '0-0-cover'.) "
-		read firstfile
-		if [ "$firstfile" = "" ]
-			then
-			firstfile="0-0-cover"
-		fi
+		# Ask whether to keep the boilerplate epub mathjax directory
+		echo "Include mathjax? Enter y for yes (or enter for no)."
+		read epubmathjax
 		# Ask the user to add any extra Jekyll config files, e.g. _config.myconfig.yml
 		echo -n "
 Any extra config files?
@@ -280,35 +276,137 @@ Enter filenames (including any relative path), comma separated, no spaces. E.g.
 _configs/_config.myconfig.yml
 If not, just hit return."
 		read config
-		# Ask whether we're processing MathJax, to know whether to send the HTML via PhantomJS
-		echo "Does this book use MathJax? If no, hit enter. If yes, type any key then enter."
-		read epubmathjax
 		# We're going to let users run this over and over by pressing enter
 		repeat=""
 		while [ "$repeat" = "" ]
 		do
 			# let the user know we're on it!
 			echo "Generating HTML..."
-			# ...and run Jekyll to build new HTML, enabling MathJax if necessary
-			if [ "$epubmathjax" = "" ]
-				then
-				bundle exec jekyll build --config="_config.yml,_configs/_config.epub.yml,$config"
-			else
+			# ...and run Jekyll to build new HTML
+			if [ $epubmathjax = "y" ]; then
 				bundle exec jekyll build --config="_config.yml,_configs/_config.epub.yml,_configs/_config.mathjax-enabled.yml,$config"
-			fi
-			# Navigate to the relevant text folder...
-			if [ "$epubsubdirectory" = "" ]
-				then
-				cd _site/$bookfolder/text
 			else
-				cd _site/$bookfolder/text/$epubsubdirectory
+				bundle exec jekyll build --config="_config.yml,_configs/_config.epub.yml,$config"
 			fi
-			# ...and open the first epub file there.
-			# Let the user know we're now going to open Sigil.
-			# NOTE: Sigil is not well supported on Linux, so many users won't have it.
-			echo "Opening Sigil..."
-			sigil "$firstfile.html"
-			# Open file browser to see epub-ready HTML files
+			# Now to assemble the epub
+			echo "Assembling epub..."
+			# Check if there are fonts to include
+			echo "Checking for fonts to include..."
+			epubfonts=""
+			countttf=`ls -1 fonts/*.ttf 2>/dev/null | wc -l`
+			if [ $countttf != 0 ]; then 
+				epubfonts="y"
+			fi
+			countotf=`ls -1 fonts/*.otf 2>/dev/null | wc -l`
+			if [ $countotf != 0 ]; then 
+				epubfonts="y"
+			fi
+			countwoff=`ls -1 fonts/*.woff 2>/dev/null | wc -l`
+			if [ $countwoff != 0 ]; then 
+				epubfonts="y"
+			fi
+			countwoff2=`ls -1 fonts/*.woff2 2>/dev/null | wc -l`
+			if [ $countwoff2 != 0 ]; then 
+				epubfonts="y"
+			fi
+			# Check if there are scripts to include
+			echo "Checking for scripts to include..."
+			epubscripts=""
+			countjs=`ls -1 js/*.js 2>/dev/null | wc -l`
+			if [ $countjs != 0 ]; then 
+				epubscripts="y"
+			fi
+			# Copy text (files in file-list only), images, fonts, styles and package.opf to epub
+			cd _site/"$bookfolder"
+			if [ "$epubsubdirectory" = "" ]; then
+				mkdir ../epub/text && cd text && cp `cat file-list` ../../epub/text/
+				cd ..
+			else
+				mkdir ../epub/text && cd $epubsubdirectory/text && cp `cat file-list` ../../../epub/text/
+				cd .. && cd ..
+			fi
+			if [ -d images ]; then
+				mkdir ../epub/images && cp -a images/. ../epub/images/
+			fi
+			if [ "$epubfonts" = "y" ]; then
+				mkdir ../epub/fonts && cp -a fonts/. ../epub/fonts/
+			fi
+			if [ -d styles ]; then
+				mkdir ../epub/styles && cp -a styles/. ../epub/styles/
+			fi
+			if [ "$epubmathjax" = "y" ]; then
+				mkdir ../epub/mathjax && cp -a ../assets/js/mathjax/. ../epub/mathjax/
+			fi
+			if [ "$epubscripts" = "y" ]; then
+				mkdir ../epub/js && cp -a js/. ../epub/js/
+			fi
+			if [ -e package.opf ]; then
+				cp package.opf ../epub/package.opf
+			fi
+		    # First, though, if they exist, remove previous .zip and .epub files that we will replace.
+			echo "Removing previous zips or epubs..."
+			if [ -e "$location/_output/$bookfolder.zip" ]; then
+				rm "$location/_output/$bookfolder.zip"
+			fi
+			if [ -e "$location/_output/$bookfolder.epub" ]; then
+				rm "$location/_output/$bookfolder.epub"
+			fi
+			# Go into _site/epub to zip it to _output
+			cd ../epub
+			# First, though, remove the fonts folder if we dont' want it
+			if [ "$epubfonts" = "" ]; then
+				rm -r fonts
+			fi
+			# And remove the mathjax dir if we don't need it
+			if [ "$epubmathjax" = "" ]; then
+				rm -r mathjax
+			fi
+			# Now to compress the epub files
+			echo "Compressing epub..."
+			# Add the mimetype first, with no compression and no extra fields (-X)
+			zip --compression-method store -0 -X --quiet "$location/_output/$bookfolder.zip" mimetype
+			if [ -d images ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "images"
+			fi
+			if [ -d fonts ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "fonts"
+			fi
+			if [ -d styles ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "styles"
+			fi
+			if [ -d text ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "text"
+			fi
+			if [ -d mathjax ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "mathjax"
+			fi
+			if [ -d js ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "js"
+			fi
+			if [ -d META-INF ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" META-INF
+			fi
+			if [ -e package.opf ]; then
+				zip --quiet "$location/_output/$bookfolder.zip" package.opf
+			fi
+	    	# Change file extension .zip to .epub
+    		cd $location/_output
+    		if [ -e "$bookfolder".zip ]; then
+				mv "$bookfolder".zip "$bookfolder".epub
+			fi
+			echo "Epub created!"
+			# Validation
+			echo "To run validation now, enter the path to the EpubCheck folder on your machine. E.g. /usr/bin/local/epubcheck-4.0.1"
+			echo "Or hit enter to skip EpubCheck validation."
+			echo "(You can get EpubCheck from https://github.com/IDPF/epubcheck/releases)"
+			read pathtoepubcheck
+			if [ "$pathtoepubcheck" = "" ]; then
+				echo "Okay, skipping EpubCheck. Try http://validator.idpf.org to validate separately."
+			else
+				java -jar "$pathtoepubcheck"/epubcheck.jar "$bookfolder".epub
+			fi
+			# Open file browser to see epub
+			# (for OSX, this is open, not xdg-open)
 			xdg-open .
 			# Navigate back to where we started
 			cd "$location"
@@ -415,6 +513,7 @@ If not, just hit return."
 			# Tell the user we're done
 			echo Done! Opening file explorer...
 			# Open file explorer to show the docx files
+			# (for OSX, this is open, not xdg-open)
 			xdg-open .
 			# Navigate back to where we began.
 			cd "$location"